Use HASH_BLOBS hashing method.
Up to pg13, if no hashing method is provided hash_create would default to
HASH_STRINGS.  Starting with pg14, a default has to be provided, and
chosing HASH_STRINGS will error out if the keysize isn't at least 8B, which
isn't the case here.

So always explicitly use HASH_BLOBS for consistency, and as it should be
slightly more efficient.